+++ .isort.cfg
@@ -1,3 +1,5 @@
 [settings]
 profile=black
 known_third_party=wandb,comet_ml
+known_local_folder=src,tests
+known_tests=tests.hf_offline_utils
