+++ tests/e2e/test_deepseekv3.py
@@ -7,13 +7,14 @@ import os
 from pathlib import Path
 
 import pytest
-from utils import enable_hf_offline
 
 from axolotl.cli.args import TrainerCliArgs
 from axolotl.common.datasets import load_datasets
 from axolotl.train import train
 from axolotl.utils.config import normalize_config, validate_config
 from axolotl.utils.dict import DictDefault
+
+from tests.hf_offline_utils import enable_hf_offline
 
 LOG = logging.getLogger("axolotl.tests.e2e")
 os.environ["WANDB_DISABLED"] = "true"
