+++ tests/prompt_strategies/conftest.py
@@ -5,10 +5,11 @@ shared fixtures for prompt strategies tests
 import pytest
 from datasets import Dataset
 from transformers import AutoTokenizer
-from utils import enable_hf_offline
 
 from axolotl.prompt_strategies.jinja_template_analyzer import JinjaTemplateAnalyzer
 from axolotl.utils.chat_templates import _CHAT_TEMPLATES
+
+from tests.hf_offline_utils import enable_hf_offline
 
 
 @pytest.fixture(name="assistant_dataset")
