+++ tests/prompt_strategies/test_dpo_chat_templates.py
@@ -7,10 +7,11 @@ import unittest
 import pytest
 from datasets import Dataset
 from transformers import AutoTokenizer
-from utils import enable_hf_offline
 
 from axolotl.prompt_strategies.dpo.chat_template import default
 from axolotl.utils.dict import DictDefault
+
+from tests.hf_offline_utils import enable_hf_offline
 
 
 @pytest.fixture(name="assistant_dataset")
